+++ assets/v030/main.js
@@ -1,8 +1,8 @@
 /*
  * @Author: laixi
  * @Date:   2017-03-14 11:34:34
- * @Last Modified by:   Xavier Yin
- * @Last Modified time: 2017-05-15 10:35:25
+ * @Last Modified by:   laixi
+ * @Last Modified time: 2017-05-15 17:33:07
  */
 import _ from 'underscore';
 import Attributes from './attributes';
@@ -10,11 +10,13 @@ import Backbone from './core';
 import Controller from './controller';
 import Events from './events';
 import Model from './model';
+import Request from './request';
 import View from './view';
 
 Backbone.Attributes = Attributes;
 Backbone.Controller = Controller;
 Backbone.Model = Model;
+Backbone.Request = Request;
 Backbone.View = View;
 
 _.extend(Backbone, Events);
@@ -25,6 +27,6 @@ _.each(['Model', 'View', 'Collection', 'Router', 'History'], function(klass) {
 
 // 扩展 Model 实例方法，增加 allChanged 和 anyChanged 方法。
 _.extend(Model.prototype, _.pick(Attributes, 'allChanged', 'anyChanged'));
-Backbone.JVERSION = '0.3.2';
+Backbone.JVERSION = '0.3.3';
 // this is invalid in es6
 module.exports = Backbone;