+++ CHANGELOG.md
@@ -12,6 +12,7 @@
   4.6. 默认混合 Attributes 对象，无需再额外扩展。
   4.7. events 与 nodes 支持 @ 语法。(v0.3.1)
   4.8. View 增加委托事件过滤机制，默认一个事件只被一个 View 处理。(v0.3.2)
+5. 新增 Backbone.Request 对象。(v0.3.3)
 
 Todos: